feat: [FE] 소켓 networking 부분 main으로 위치 수정
- Intent
